requirements.txt
grpcio>=1.22.0
grpcio-tools
joblib==0.13.2
tensorflow==1.13.1
tensorflow_hub==0.4.0
numpy==1.16.4
protobuf==3.9.0

